Refactor telemetry collecting code
- factor out all telemetry and health checking code into
  monitor_and_control_chain_health(), remove duplicate code from
  A1_bench_scanwork()

- introduce distinct intervals for (arbitrary values for now, to be tweaked):
  - temperature (TEMP_UPDATE_INT_MS)
  - voltage (replace VOLTAGE_UPDATE_INT with VOLTAGE_UPDATE_INT_MS in
    milliseconds)
  - telemetry submission period

- move check for disabled chips to
  monitor_and_control_chain_health(). Code is still be
  simplifed/refactored due to unnecessary A1Pll[1-6] multiplexing

- change signature of get_current_ms() - use unsigned long and use
  correct time comparing logic that is not sensitive to overflow in
  monitor_and_control_chain_health().

- change A1_chip::cooldown_begin to unsigned long and fix broken time
  comparison

- add new fields in A1_chip structure (last_voltage_time,
  last_telemetry_time) to support new monitoring algorithm

+++ A5_inno.h
@@ -49,7 +49,7 @@ void prechain_detect(struct A1_chain *a1, int idxpll);
 int chain_detect(struct A1_chain *a1);
 bool abort_work(struct A1_chain *a1);
 
-int get_current_ms(void);
+unsigned long get_current_ms(void);
 bool is_chip_disabled(struct A1_chain *a1, uint8_t chip_id);
 void disable_chip(struct A1_chain *a1, uint8_t chip_id);
